+++ NEWS.md
@@ -5,6 +5,7 @@
 - Added dtm_align
 - Added dtm_sample
 - Added document_term_matrix.matrix
+- dtm_cbind and dtm_rbind allow to pass more than 2 sparse matrices
 - cbind_morphological gains argument which to specify which morphological features to extract
 - txt_count now returns NA when NA is provided instead of an error
 - txt_contains now returns NA when NA is provided instead of FALSE, unless value is set to TRUE